gcs-migrator/requirements.txt
altgraph==0.17.3
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
google-api-core==2.11.0
google-auth==2.16.0
google-cloud-core==2.3.2
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
grpcio==1.51.1
grpcio-tools==1.51.1
idna==3.4
joblib==1.2.0
macholib==1.16.2
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyelftools==0.29
pyinstaller==5.7.0
pyinstaller-hooks-contrib==2022.15
python-json-logger==2.0.4
requests==2.28.2
rsa==4.9
SCons==4.4.0
six==1.16.0
staticx==0.13.8
urllib3==1.26.14
